New alpine non-root docker image build
Old Docker images used 'go run', which does a build of the code every
time it is run and requires internet access to run. The docker image was
also bloated as it was a full ubuntu image with golang compiler
installed.

New docker image is alpine based and includes statically-built binaries
that are stored in /app.

Additionally, the daemons are all run as non-root user id an 'telemetry'
user.